refactor(coins/test): move data test constants to fakes package
