--- data/src/test/java/com/hxl/data/DataTestConstants.java
@@ -1,16 +1,0 @@
-package com.hxl.data;
-
-public class DataTestConstants {
-    // Keys
-    public static final String KEY_INT = "int";
-    public static final String KEY_BOOL = "bool";
-    public static final String KEY_STR = "str";
-    // Values
-    public static final int VAL_INT = 123;
-    public static final boolean VAL_BOOL = true;
-    public static final String VAL_STR = "test-str";
-    // Default values
-    public static final int DEF_INT = 321;
-    public static final boolean DEF_BOOL = false;
-    public static final String DEF_STR = "default";
-}
